+++ indico/modules/users/client/js/PersonalDataForm.jsx
@@ -36,6 +36,8 @@ function PersonalDataForm({
   currentAffiliation,
   titles,
   syncedValues,
+  lockedFields,
+  lockedFieldMessage,
   hasPredefinedAffiliations,
 }) {
   const userIdArgs = userId !== null ? {user_id: userId} : {};
@@ -87,12 +89,16 @@ function PersonalDataForm({
                 label={Translate.string('First name')}
                 required
                 syncedValues={syncedValues}
+                lockedFields={lockedFields}
+                lockedFieldMessage={lockedFieldMessage}
               />
               <SyncedFinalInput
                 name="last_name"
                 label={Translate.string('Last name')}
                 required
                 syncedValues={syncedValues}
+                lockedFields={lockedFields}
+                lockedFieldMessage={lockedFieldMessage}
               />
             </Form.Group>
             {hasPredefinedAffiliations ? (
@@ -101,28 +107,38 @@ function PersonalDataForm({
                 syncName="affiliation"
                 currentAffiliation={currentAffiliation}
                 syncedValues={syncedValues}
+                lockedFields={lockedFields}
+                lockedFieldMessage={lockedFieldMessage}
               />
             ) : (
               <SyncedFinalInput
                 name="affiliation"
                 label={Translate.string('Affiliation')}
                 syncedValues={syncedValues}
+                lockedFields={lockedFields}
+                lockedFieldMessage={lockedFieldMessage}
               />
             )}
             <SyncedFinalTextArea
               name="address"
               label={Translate.string('Address')}
               syncedValues={syncedValues}
+              lockedFields={lockedFields}
+              lockedFieldMessage={lockedFieldMessage}
             />
             <SyncedFinalInput
               name="phone"
               label={Translate.string('Phone number')}
               syncedValues={syncedValues}
+              lockedFields={lockedFields}
+              lockedFieldMessage={lockedFieldMessage}
             />
             <SyncedFinalInput
               name="email"
               label={Translate.string('Email address')}
               syncedValues={syncedValues}
+              lockedFields={lockedFields}
+              lockedFieldMessage={lockedFieldMessage}
               readOnly
             >
               <Translate>
@@ -152,6 +168,8 @@ PersonalDataForm.propTypes = {
     })
   ).isRequired,
   syncedValues: PropTypes.objectOf(PropTypes.string).isRequired,
+  lockedFields: PropTypes.arrayOf(PropTypes.string).isRequired,
+  lockedFieldMessage: PropTypes.string.isRequired,
   hasPredefinedAffiliations: PropTypes.bool.isRequired,
 };
 
@@ -166,6 +184,8 @@ window.setupPersonalDataForm = function setupPersonalDataForm(
   currentAffiliation,
   titles,
   syncedValues,
+  lockedFields,
+  lockedFieldMessage,
   hasPredefinedAffiliations
 ) {
   document.addEventListener('DOMContentLoaded', () => {
@@ -176,6 +196,8 @@ window.setupPersonalDataForm = function setupPersonalDataForm(
         currentAffiliation={currentAffiliation}
         titles={titles}
         syncedValues={syncedValues}
+        lockedFields={lockedFields}
+        lockedFieldMessage={lockedFieldMessage}
         hasPredefinedAffiliations={hasPredefinedAffiliations}
       />,
       document.querySelector('#personal-details-form-container')