+++ indico/web/client/js/react/components/syncedInputs.jsx
@@ -10,7 +10,7 @@ import searchAffiliationURL from 'indico-url:users.api_affiliations';
 import PropTypes from 'prop-types';
 import React, {useState} from 'react';
 import {useField, useForm} from 'react-final-form';
-import {Header} from 'semantic-ui-react';
+import {Form, Header, Popup} from 'semantic-ui-react';
 
 import {FinalComboDropdown, FinalDropdown, FinalInput, FinalTextArea} from 'indico/react/forms';
 import {Translate} from 'indico/react/i18n';
@@ -26,6 +26,8 @@ function SyncedFinalField({
   syncName,
   as: FieldComponent,
   syncedValues,
+  lockedFields,
+  lockedFieldMessage,
   readOnly,
   required,
   processSyncedValue,
@@ -41,25 +43,28 @@ function SyncedFinalField({
   }
 
   const syncable = syncedValues[syncName] !== undefined && (!required || syncedValues[syncName]);
+  const synced = syncedFields.includes(syncName);
+  const locked = lockedFields.includes(syncName);
 
-  return (
+  const field = (
     <FieldComponent
       {...rest}
       name={name}
       styleName={syncable ? 'syncable' : ''}
-      readOnly={readOnly || (syncable && syncedFields.includes(syncName))}
+      readOnly={readOnly || (syncable && synced)}
       required={required}
       action={
         syncable
           ? {
               type: 'button',
-              active: syncedFields.includes(syncName),
+              active: synced,
+              disabled: synced && locked,
               icon: 'sync',
               toggle: true,
               className: 'ui-qtip',
               title: Translate.string('Toggle synchronization of this field'),
               onClick: () => {
-                if (syncedFields.includes(syncName)) {
+                if (synced) {
                   setSyncedFields(syncedFields.filter(x => x !== syncName));
                   form.change(name, form.getFieldState(name).initial);
                 } else {
@@ -72,6 +77,18 @@ function SyncedFinalField({
       }
     />
   );
+  if (synced && locked && lockedFieldMessage) {
+    return (
+      <Popup
+        on="hover"
+        position="bottom left"
+        content={lockedFieldMessage}
+        trigger={<Form.Field>{field}</Form.Field>}
+      />
+    );
+  } else {
+    return field;
+  }
 }
 
 SyncedFinalField.propTypes = {
@@ -79,6 +96,8 @@ SyncedFinalField.propTypes = {
   syncName: PropTypes.string,
   as: PropTypes.elementType.isRequired,
   syncedValues: PropTypes.object.isRequired,
+  lockedFields: PropTypes.arrayOf(PropTypes.string).isRequired,
+  lockedFieldMessage: PropTypes.string.isRequired,
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
   processSyncedValue: PropTypes.func,
@@ -109,6 +128,8 @@ export function SyncedFinalAffiliationDropdown({
   required,
   syncName,
   syncedValues,
+  lockedFields,
+  lockedFieldMessage,
   currentAffiliation,
 }) {
   const [_affiliationResults, setAffiliationResults] = useState([]);
@@ -167,6 +188,8 @@ export function SyncedFinalAffiliationDropdown({
       )}
       label={Translate.string('Affiliation')}
       syncedValues={syncedValues}
+      lockedFields={lockedFields}
+      lockedFieldMessage={lockedFieldMessage}
     />
   );
 }
@@ -176,6 +199,8 @@ SyncedFinalAffiliationDropdown.propTypes = {
   required: PropTypes.bool,
   syncName: PropTypes.string,
   syncedValues: PropTypes.object.isRequired,
+  lockedFields: PropTypes.arrayOf(PropTypes.string).isRequired,
+  lockedFieldMessage: PropTypes.string.isRequired,
   currentAffiliation: PropTypes.object,
 };