+++ CHANGES.rst
@@ -35,6 +35,9 @@ Improvements
 - Hide the top box with the latest files of an editable until it has been accepted
   and published (:issue:`5660`, :pr:`5665`)
 - Allow uploading files when requesting changes on the editing timeline (:pr:`5612`)
+- Add ``locked_fields`` to the identity provider settings in ``indico.conf`` to
+  prevent non-admin users from turning off their profile's personal data
+  synchronization (:pr:`5648`)
 
 Bugfixes
 ^^^^^^^^